+++ src/main/java/com/sealog/backend/domain/feature/user/controller/UserAdminController.java
@@ -1,8 +1,7 @@
 package com.sealog.backend.domain.feature.user.controller;
 
 import com.sealog.backend.domain.feature.auth.dto.AuthRequest;
-import com.sealog.backend.domain.feature.auth.service.AuthService;
-import com.sealog.backend.domain.feature.user.service.AdminUserService;
+import com.sealog.backend.domain.feature.user.service.UserService;
 import com.sealog.backend.global.response.CustomResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -15,9 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/admin/user")
 @RequiredArgsConstructor
-public class AdminUserController implements AdminUserControllerDocs {
+public class UserAdminController implements UserAdminControllerDocs {
 
-    private final AdminUserService adminUserService;
+    private final UserService userAdminService;
 
     /**
      * 회원가입
@@ -29,7 +28,7 @@ public class AdminUserController implements AdminUserControllerDocs {
     public ResponseEntity<CustomResponse<Void>> signUp(
             @Valid @RequestBody AuthRequest.SignUpRequest request
     ) {
-        adminUserService.signUp(request);
+        userAdminService.signUp(request);
         return ResponseEntity.ok(CustomResponse.success(null, "회원가입이 성공적으로 완료되었습니다."));
     }
 }
